+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Added: Schema test for the .debug.ps1 file
 * Added: Schema tests for the (helper) function files and export definitions
 * Added: Schema tests for encoding, white space and indentation character
+* Changed: Verify task now allows newer revisions for InvokeBuildHelper
 * Changed: Move asserts from Gallery to Approve task
 
 ## 1.0.0 - 2019-10-23